refactor: BatchDropRateModalをauto/manual別コンポーネントに分割
830行のモノリシックコンポーネントを3ファイルに分割:
- BatchDropRateModal.tsx: ルーター（autoMode分岐のみ）
- BatchDropRateAutoContent.tsx: 自動モード（intra_rarity_weight操作、確率プレビューをサーバー計算式と一致）
- BatchDropRateManualContent.tsx: 手動モード（既存動作維持、modifiedCardIds閾値修正）

CardManager.tsxのProps互換維持、API変更なし。
